refactor: split group command
